+++ unit-tests/cl_mean_center.cpp
@@ -52,7 +52,7 @@ BOOST_AUTO_TEST_CASE(mean_center, * but::fixture(&setup))
   auto mat_f = mfmat::cl_mat_f{400, 500};
   for (std::size_t m = 0; m < mat_f.get_row_count(); ++m)
     for (std::size_t n = 0; n < mat_f.get_col_count(); ++n)
-      mat_f[{ m, n }] = m + n;
+      mat_f[{ m, n }] = static_cast<float>(m + n);
   auto tmp_mean_f = mfmat::mean(mat_f);
   auto tmp_f = mat_f - tmp_mean_f;
   mat_f.mean_center();
@@ -71,7 +71,7 @@ BOOST_AUTO_TEST_CASE(mean_center, * but::fixture(&setup))
   auto mat_d = mfmat::cl_mat_d{700, 600};
   for (std::size_t m = 0; m < mat_d.get_row_count(); ++m)
     for (std::size_t n = 0; n < mat_d.get_col_count(); ++n)
-      mat_d[{ m, n }] = m + n;
+      mat_d[{ m, n }] = static_cast<double>(m + n);
   auto tmp_mean_d = mfmat::mean(mat_d);
   auto tmp_d = mat_d - tmp_mean_d;
   mat_d.mean_center();
